requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.29
sqlalchemy>=2.0
pydantic>=2.5
pydantic-settings>=2.0
requests>=2.31
jinja2>=3.1
python-multipart>=0.0.9
python-dotenv>=1.0
itsdangerous>=2.1
starlette>=0.36
google-auth>=2.28
google-auth-oauthlib>=1.2
google-auth-httplib2>=0.2
google-api-python-client>=2.120
httpx>=0.27
aiofiles>=23.2
